+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.54.1
 ### Features
 * Added a temporary method `ContainerConfigBuilder.WithPublicPorts` to support using `static public ports` for user services